Enum automerge_frontend::Primitive

pub enum Primitive {
    Bytes(Vec<u8>),
    Str(SmolStr),
    Int(i64),
    Uint(u64),
    F64(f64),
    Counter(i64),
    Timestamp(i64),
    Boolean(bool),
    Cursor(Cursor),
    Null,
}

A primitive value, leaf nodes of the document tree.

Variants

Bytes(Vec<u8>)

Tuple Fields

0: Vec<u8>

Str(SmolStr)

Tuple Fields

0: SmolStr

Int(i64)

Tuple Fields

0: i64

Uint(u64)

Tuple Fields

0: u64

F64(f64)

Tuple Fields

0: f64

Counter(i64)

Tuple Fields

0: i64

Timestamp(i64)

Tuple Fields

0: i64

Boolean(bool)

Tuple Fields

0: bool

Cursor(Cursor)

Tuple Fields

0: Cursor

Null

Implementations

impl Primitive

pub fn is_bytes(&self) -> bool

Return whether the Primitive is bytes.

pub fn bytes(&self) -> Option<&[u8]>

Extract the &[u8] in this Primitive if it represents bytes.

pub fn is_str(&self) -> bool

Return whether the Primitive is a string.

pub fn str(&self) -> Option<&str>

Extract the &str in this Primitive if it represents a string.

pub fn is_int(&self) -> bool

Return whether the Primitive is an int.

pub fn int(&self) -> Option<i64>

Extract the i64 in this Primitive if it represents an int.

pub fn is_uint(&self) -> bool

Return whether the Primitive is a uint.

pub fn uint(&self) -> Option<u64>

Extract the u64 in this Primitive if it represents a uint.

pub fn is_f64(&self) -> bool

Return whether the Primitive is a float.

pub fn f64(&self) -> Option<f64>

Extract the f64 in this Primitive if it represents a float.

pub fn is_counter(&self) -> bool

Return whether the Primitive is a counter.

pub fn counter(&self) -> Option<i64>

Extract the i64 in this Primitive if it represents a counter.

pub fn is_timestamp(&self) -> bool

Return whether the Primitive is a timestamp.

pub fn timestamp(&self) -> Option<i64>

Extract the i64 in this Primitive if it represents a timestamp.

pub fn is_boolean(&self) -> bool

Return whether the Primitive is a boolean.

pub fn boolean(&self) -> Option<bool>

Extract the bool in this Primitive if it represents a boolean.

pub fn is_cursor(&self) -> bool

Return whether the Primitive is a cursor.

pub fn cursor(&self) -> Option<&Cursor>

Extract the Cursor in this Primitive if it represents a cursor.

pub fn is_null(&self) -> bool

Return whether the Primitive is null.