Enum automerge_frontend::Value

pub enum Value {
    Map(HashMap<SmolStr, Value>),
    Table(HashMap<SmolStr, Value>),
    List(Vec<Value>),
    Text(Vec<SmolStr>),
    Primitive(Primitive),
}

A composite value, composing maps, tables, lists, text and primitives.

A Value is the general container type for objects in the document tree.

Variants

Map(HashMap<SmolStr, Value>)

Tuple Fields

0: HashMap<SmolStr, Value>

A mapping from string keys to values.

Table(HashMap<SmolStr, Value>)

Tuple Fields

0: HashMap<SmolStr, Value>

A mapping from unique IDs to values.

List(Vec<Value>)

Tuple Fields

0: Vec<Value>

An ordered sequence of values.

Text(Vec<SmolStr>)

Tuple Fields

0: Vec<SmolStr>

An ordered sequence of grapheme clusters.

Primitive(Primitive)

Tuple Fields

0: Primitive

A primitive value.

Implementations

impl Value

pub fn is_map(&self) -> bool

Return whether the Value is a map.

pub fn map(&self) -> Option<&HashMap<SmolStr, Value>>

Extract the inner map in this Value if it represents a map.

pub fn is_table(&self) -> bool

Return whether the Value is a table.

pub fn table(&self) -> Option<&HashMap<SmolStr, Value>>

Extract the inner map in this Value if it represents a table.

pub fn is_list(&self) -> bool

Return whether the Value is a list.

pub fn list(&self) -> Option<&[Value]>

Extract the elements in this Value if it represents a list.

pub fn is_text(&self) -> bool

Return whether the Value is text.

pub fn text(&self) -> Option<&[SmolStr]>

Extract the graphemes in this Value if it represents text.

pub fn is_primitive(&self) -> bool

Return whether the Value is a primitive.

pub fn primitive(&self) -> Option<&Primitive>

Extract the Primitive in this Value if it represents a primitive.

pub fn from_json(json: &Value) -> Value

Convert a JSON object into a Value.

pub fn to_json(&self) -> Value

Convert this Value into a JSON object.

pub fn get_value(&self, path: Path) -> Option<Cow<'_, Self>>

Get the Value at the given path, if one exists.