Struct automerge_frontend::Frontend

pub struct Frontend {
    pub actor_id: ActorId,
    pub seq: u64,
    // some fields omitted
}

Fields
actor_id: ActorIdseq: u64
Implementations

impl Frontend

pub fn new() -> Self

pub fn new_with_actor_id(actor_id: &[u8]) -> Self

pub fn new_with_timestamper(t: Box<dyn Fn() -> Option<i64>>) -> Self

pub fn new_with_timestamper_and_actor_id(
t: Box<dyn Fn() -> Option<i64>>,
actor_id: &[u8]
) -> Self

pub fn new_with_initial_state(
initial_state: Value
) -> Result<(Self, Change), InvalidInitialStateError>

pub fn state(&mut self) -> &Value

pub fn change<F, O, E>(
&mut self,
message: Option<String>,
change_closure: F
) -> Result<(O, Option<Change>), E> where
E: Error,
F: FnOnce(&mut dyn MutableDocument) -> Result<O, E>,

pub fn apply_patch(&mut self, patch: Patch) -> Result<(), InvalidPatch>

pub fn get_object_id(&self, path: &Path) -> Option<ObjectId>

pub fn in_flight_requests(&self) -> Vec<u64>

pub fn get_conflicts(&self, path: &Path) -> Option<HashMap<OpId, Value>>

Gets the set of values for path, returns None if the path does not exist

pub fn get_value(&self, path: &Path) -> Option<Value>

Returns the value given by path, if it exists